pyproject.toml
[project]
name = "Snowcode"
version = "0.3.1"
description = "Local RAG-powered CLI tool for chatting with your codebase"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Snowcode" }]
keywords = ["rag", "llm", "cli", "codebase", "search", "ai", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "numpy>=1.24",
    "sentence-transformers>=3.0",
    "rich>=13.0",
    "pathspec>=0.12",
    "openai>=1.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/HUAYUE1024/codechat"
Repository = "https://github.com/HUAYUE1024/codechat"
Issues = "https://github.com/HUAYUE1024/codechat/issues"

[project.optional-dependencies]
ast = [
    "tree-sitter>=0.22",
    "tree-sitter-languages>=1.10",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
multimodal = [
    "Pillow>=10.0",        # Image processing
    "pytesseract>=0.3",    # OCR text extraction
    "PyMuPDF>=1.23",       # PDF reading (fitz)
    "python-docx>=1.1",    # Word document reading
    "openpyxl>=3.1",       # Excel spreadsheet reading
    "beautifulsoup4>=4.12", # HTML text extraction
]
scientific = [
    "netCDF4>=1.6",        # NetCDF file reading
    "xarray>=2023.0",      # Multi-dimensional data analysis
]

[project.scripts]
snowcode = "codechat.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["codechat*"]

[tool.setuptools.package-data]
codechat = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=codechat --cov-report=term-missing"
